refactor: fixed TS support for mock
