+++ example/jestSetup.js
@@ -1,5 +1,5 @@
 import 'react-native-gesture-handler/jestSetup';
 
 jest.mock('react-native-keyboard-controller', () =>
-  require('react-native-keyboard-controller/jest/mock')
+  require('react-native-keyboard-controller/jest')
 );
