+++ example/src/screens/Examples/KeyboardAnimation/spec/index.spec.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import { render } from '@testing-library/react-native';
 
-import { setKeyboardPosition } from 'react-native-keyboard-controller/jest/mock';
+import { setKeyboardPosition } from 'react-native-keyboard-controller/jest';
 
 import KeyboardAnimation from '../index';
-
-jest.mock('react-native/Libraries/Animated/NativeAnimatedHelper');
 
 describe('unit test sample', () => {
   it('should match to snapshot', () => {
