+++ jest/index.d.ts
@@ -1,0 +1,4 @@
+declare module 'react-native-keyboard-controller/jest' {
+  function setKeyboardPosition(height: number);
+  function setKeyboardVisibleHeight(height: number);
+}
